# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CHELSY Restaurant API",
        "description": "API REST pour l'application mobile CHELSY Restaurant",
        "contact": {
            "email": "contact@chelsy-restaurant.bj"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Serveur API"
        }
    ],
    "paths": {
...
        "/api/v1/payments/confirm-stripe": {
            "post": {
                "tags": [
                    "Paiements"
                ],
                "summary": "Confirmer un paiement Stripe",
                "operationId": "f5235c0a575d1bc6236519f37f5d34ba",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "order_id",
                                    "payment_intent_id"
                                ],
                                "properties": {
                                    "order_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "payment_intent_id": {
                                        "type": "string",
                                        "example": "pi_xxxxxxxxxxxxx"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Paiement confirmé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Paiement confirmé avec succès"
                                        },
                                        "data": {
                                            "properties": {
                                                "order": {
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation ou paiement non confirmé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/webhooks/stripe": {
            "post": {
                "tags": [
                    "Paiements"
                ],
                "summary": "Webhook Stripe pour les notifications de paiement",
                "operationId": "40376bfd2b826022a0cfed17692535c2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "type": {
                                        "type": "string",
                                        "example": "payment_intent.succeeded"
                                    },
                                    "data": {
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Webhook reçu",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "received": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Authentification par token Sanctum",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}